Add files via upload
In this project, I created a game where a user has to guess a color that the program generated randomly.  This project contains arrays, Math.random()/Math.floor(), do/while and if/else if loops, and iterations are shown to the user. Once the user correctly guesses the color, the background of the web page is changed to that color.